getting closer to finishing get_macro endpoint

+++ models/diet_plan.go
@@ -10,8 +10,8 @@ type Diet_Plan struct {
 	Main_Image_Link    string    `json:"main_image_link"`
 	Background_Color   string    `json:"background_color"`
 	Diet_Plan_Desc     string    `json:"diet_plan_desc"`
-	Calorie_Percentage int16     `json:"calorie_percentage"`
-	Protein_Percentage int16     `json:"protein_percentage"`
-	Fats_Percentage    int16     `json:"fats_percentage"`
-	Carbs_Percentage   int16     `json:"carbs_percentage"`
+	Calorie_Percentage int       `json:"calorie_percentage"`
+	Protein_Percentage int       `json:"protein_percentage"`
+	Fats_Percentage    int       `json:"fats_percentage"`
+	Carbs_Percentage   int       `json:"carbs_percentage"`
 }
